refactor(fetcher): split error handling into separate file
- Move ExchangeError and FetchError classes to new fetcherError.ts file
- Update import statements and error handling in fetcher.ts
- Add new unit tests for FetchError in fetcherError.test.ts
- Remove ExchangeError tests from fetchExchange.test.ts
