+++ packages/fetcher/src/fetchExchange.ts
@@ -15,28 +15,6 @@ import { Fetcher } from './fetcher';
 import { FetchRequest } from './fetchRequest';
 
 /**
- * Custom error class for FetchExchange related errors.
- *
- * This error is thrown when there are issues with the HTTP exchange process,
- * such as when a request fails and no response is generated.
- */
-export class ExchangeError extends Error {
-  constructor(
-    public readonly exchange: FetchExchange,
-  ) {
-    const errorMessage = exchange.error?.message ||
-      exchange.response?.statusText ||
-      'Unknown error occurred during exchange';
-    super(errorMessage);
-    this.name = 'ExchangeError';
-    if (exchange.error?.stack) {
-      this.stack = exchange.error.stack;
-    }
-    Object.setPrototypeOf(this, ExchangeError.prototype);
-  }
-}
-
-/**
  * Container for HTTP request/response data that flows through the interceptor chain.
  *
  * Represents the complete exchange object that flows through the interceptor chain.
